14-08-2024

+++ database/migrations/2024_08_08_030337_update_coffees_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('coffees', function (Blueprint $table) {
             $table->string('image')->after('description')->nullable(true);
-            $table->float('rate')->after('image')->default(0);
+            $table->decimal('rate')->after('image')->default(0);
             $table->string('review')->after('rate')->default('');
         });
     }
